feat(markdowntext): add image height constraint option
This commit introduces an optional `imageHeight` parameter to the `MarkdownText` composable and related internal components.

When `imageHeight` is specified, images within the Markdown content will be constrained to that height. If `imageHeight` is `null` (the default), images will not have a height constraint.

This allows for better control over image display in Markdown, particularly in the `MarkdownContentRenderer` where a default height of 512 is now used (with a TODO to make it configurable in settings).

The following changes were made:
- Added `imageHeight: Int?` parameter to `MarkdownText`.
- Propagated `imageHeight` to `MarkdownRender.create` and `ImagesPlugin.create`.
- In `ImagesPlugin`, if `imageHeight` is provided, `FixedHeightImageSpanFactory` is used; otherwise, the default `ImageSpanFactory` is used.