fix(network): ensure default proxy is Proxy.NO_PROXY
If `DxrSettings.Models.httpProxy` is null or its `enabled` property is false, or if creating the `InetSocketAddress` fails, the OkHttpClient will now correctly default to using `Proxy.NO_PROXY`. Previously, it would default to null, which could lead to unexpected behavior.

We didn't find this bug because of the lack of a proxy-free environment.